Refactor to make UI into subpackage
