+++ src/main/java/duke/ui/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.ui;
 import java.util.Scanner;
 
 /**
